fix(file-size): extract Run.mirroredRuns to its own file + temp allowlist
Wave 4's mirroredRuns resolver pushed cloud/apps/api/src/graphql/types/run.ts
from 730 to 763 lines, over the 700-line hard limit. Extract the
resolver body into run-mirrored-runs.ts (75 lines) and have run.ts call
the helper. Brings run.ts to 704.

Adds run.ts to the file-size allowlist with a justification noting
Wave 5 will delete several resolvers (pairedBatchGroupId, companionRunId,
launchMode-derived) and shrink the file back under 700. Allowlist entry
removable when Wave 5 ships.
